Module Swhids.Parser

val parse : Stdlib.Lexing.lexbuf -> (Lang.identifier, string) Stdlib.result

parse buf parses a swhid from the buf Stdlib.Lexing.lexbuf.

val from_string : string -> (Lang.identifier, string) Stdlib.result

from_string s parses a swhid from string s.

val from_channel : Stdlib.in_channel -> (Lang.identifier, string) Stdlib.result

from_channel c parses a swhid from channel c

val from_file : string -> (Lang.identifier, string) Stdlib.result

from_file f parses a swhid from the file of name f.