Module Swhids__Compute

val content_identifier : string -> Swhids.Lang.identifier option

Compute the software heritage identifier for a given content

val directory_identifier : (string * int * Stdlib.String.t * string) list -> Swhids.Lang.identifier option

Compute the software heritage identifier for a given directory

val release_identifier : target:string -> Swhids.Lang.object_type -> name:string -> author:string option -> date:(int * int * int * bool) option -> message:string option -> Swhids.Lang.identifier option

Compute the software heritage identifier for a given release

val revision_identifier : string -> string list -> author:string option -> author_date:(int * int * int * bool) option -> committer:string option -> committer_date:(int * int * int * bool) option -> (string * string) array -> string option -> Swhids.Lang.identifier option

Compute the software heritage identifier for a given revision

val snapshot_identifier : (Stdlib.String.t * (string * string) option) list -> Swhids.Lang.identifier option

Compute the software heritage identifier for a given snapshot