Module Swhids__Download

val instance : string Stdlib.ref

The base URL of the software heritage instance used, defaults to https://archive.softwareheritage.org.

val content : ?⁠hash_type:string -> Swhids.Lang.object_id -> (string, string) Stdlib.result

For a given content identifier, compute an URL from which the content can be downloaded.

val directory : Swhids.Lang.object_id -> (string, string) Stdlib.result

For a given directory identifier, compute an URL from which the directory can be downloaded.

val revision : Swhids.Lang.object_id -> (string, string) Stdlib.result

For a given revision identifier, compute an URL from which the revision can be downloaded.

val release : Swhids.Lang.object_id -> (string, string) Stdlib.result

For a given release identifier, compute an URL from which the release can be downloaded.

val snapshot : Swhids.Lang.object_id -> ((string, string) Stdlib.result list, string) Stdlib.result

For a given snapshot identifier, compute a list of URL from which the snapshot's branches can be downloaded.

val any : Swhids.Lang.identifier -> ((string, string) Stdlib.result list, string) Stdlib.Result.t

For any object identifier, compute a list of URLs from which the object can be downloaded. For all kind of object, the list should contain a signe URL except for snapshot objects which may lead to a list of many URLs (one URL per branch).