Module Swhids.Compute

val content_identifier : string -> Lang.identifier option

content_identifier s computes the swhid for the s content. s is the raw content of a file as a string.

E.g. content_identifier "_build\n" is the swhid of this library's .gitignore file.

val directory_identifier : Lang.directory_entry list -> Lang.identifier option

directory_identifier entries compute the swhid for the entries directory. entries is a list of Lang.directory_entry where each element points to another object (usually a file content or a sub-directory).

E.g. directory_identifier [ { typ = "file" ; permissions = 33188 ; name = "README" ; target = "37ec8ea2110c0b7a32fbb0e872f6e7debbf95e21" }] is the swhid of a directory which has a single file README with permissions 33188 and whose core identifier from content_identifier is 37ec8ea2110c0b7a32fbb0e872f6e7debbf95e21.

val release_identifier : target:string -> Lang.object_type -> name:string -> author:string option -> date:(int * int * bool) option -> message:string option -> Lang.identifier option

Compute the software heritage identifier for a given release

val revision_identifier : string -> string list -> author:string -> author_date:(int * int * bool) option -> committer:string -> committer_date:(int * int * bool) option -> (string * string) array -> string option -> Lang.identifier option

Compute the software heritage identifier for a given revision

val snapshot_identifier : (Stdlib.String.t * (string * string) option) list -> Lang.identifier option

Compute the software heritage identifier for a given snapshot